% Generated by roxygen2 (4.0.2): do not edit by hand
\name{icbparmplot}
\alias{icbparmplot}
\title{Chargebalance vs. pcode scatter plot}
\usage{
icbparmplot(qw.data, new.threshold, icbparm.site.selection, icbparm.plotparm)
}
\arguments{
\item{qw.data}{A qw.data object generated from NWISPullR}

\item{new.threshold}{The threshold value in seconds from current system time for "new" data.}

\item{icbparm.site.selection}{A character vector of site IDs to plot}

\item{icbparm.plotparm}{A character vector of parameters to plot}
}
\description{
Takes output data object from NWISPullR and prints a plot of charge balance vs. a parameter. Requires charge balance = TRUE in NWISPullR
}

