% Generated by roxygen2 (4.0.2): do not edit by hand
\name{icbplot}
\alias{icbplot}
\title{Chargebalance timeseries plot}
\usage{
icbplot(qw.data, new.threshold, icb.site.selection, icb.begin.date.slider,
  icb.end.date.slider, icb.show.smooth)
}
\arguments{
\item{qw.data}{A qw.data object generated from NWISPullR}

\item{new.threshold}{The threshold value in seconds from current system time for "new" data.}

\item{icb.site.selection}{A character vector of site IDs to plot}

\item{icb.begin.date.slider}{Character string of begining date range to plot (yyyy-mm-dd)}

\item{icb.end.date.slider}{Character string of ending date range to plot (yyyy-mm-dd)}

\item{icb.show.smooth}{Add a loess smooth to plot}
}
\description{
Takes output data object from NWISPullR and prints a plot of charge balance vs time. Requires charge balance = TRUE in NWISPullR
}

