% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ionbalance}
\alias{ionbalance}
\title{Ion balance}
\usage{
ionbalance(qw.data)
}
\arguments{
\item{qw.data}{A qw.data object generated from NWISPullR}
}
\description{
Takes output data object from NWISPullR and calculates a QWData standard ion charge balance.
}

