% Generated by roxygen2 (4.0.2): do not edit by hand
\name{iparmbox}
\alias{iparmbox}
\title{Parameter boxplot}
\usage{
iparmbox(qw.data, new.threshold, iparmbox.site.selection, iparmbox.plotparm,
  iparmbox.show.points, iparmbox.log.scale)
}
\arguments{
\item{qw.data}{A qw.data object generated from NWISPullR}

\item{new.threshold}{The threshold value in seconds from current system time for "new" data.}

\item{iparmbox.site.selection}{A character vector of site IDs to plot}

\item{iparmbox.plotparm}{A character vector of pcodes to plot}

\item{iparmbox.show.points}{Show samples points}

\item{iparmbox.log.scale}{Plot y axis on a log scale}
}
\description{
Takes output data object from NWISPullR and prints a boxplot of parameters
}

