% Generated by roxygen2 (4.0.2): do not edit by hand
\name{iscsumplot}
\alias{iscsumplot}
\title{Sum anions vs conductance plot}
\usage{
iscsumplot(qw.data, new.threshold, iscsum.site.selection, iscsum.plotparm)
}
\arguments{
\item{qw.data}{A qw.data object generated from NWISPullR}

\item{new.threshold}{The threshold value in seconds from current system time for "new" data.}

\item{iscsum.site.selection}{A character vector of site IDs to plot}

\item{iscsum.plotparm}{A character string of parameters to plot. Options are "Cations", "Anions", "Both"}
}
\description{
Takes output data object from NWISPullR and prints a plot of sum ions vs. conductance. Requires charge balance = TRUE in NWISPullR
}

