% Generated by roxygen2 (4.0.2): do not edit by hand
\name{iseasonalbox}
\alias{iseasonalbox}
\title{Seasonal boxplot}
\usage{
iseasonalbox(qw.data, new.threshold, iseasonalbox.site.selection,
  iseasonalbox.plotparm, iseasonalbox.show.points, iseasonalbox.log.scale)
}
\arguments{
\item{qw.data}{A qw.data object generated from NWISPullR}

\item{new.threshold}{The threshold value in seconds from current system time for "new" data.}

\item{iseasonalbox.site.selection}{A character vector of site IDs to plot}

\item{iseasonalbox.plotparm}{A character vector of parameters to plot}

\item{iseasonalbox.show.points}{A logical to show points on plot}

\item{iseasonalbox.log.scale}{A logical to plot y axis on log scale}
}
\description{
Takes output data object from NWISPullR and prints a boxplot of a parameter by month.
}

