% Generated by roxygen2 (4.0.2): do not edit by hand
\name{itsplot}
\alias{itsplot}
\title{Parameter timeseries plot}
\usage{
itsplot(qw.data, new.threshold, its.site.selection, its.plotparm,
  its.begin.date.slider, its.end.date.slider, its.show.q, its.show.smooth)
}
\arguments{
\item{qw.data}{A qw.data object generated from NWISPullR}

\item{new.threshold}{The threshold value in seconds from current system time for "new" data.}

\item{its.site.selection}{A character vector of site IDs to plot}

\item{its.begin.date.slider}{Character string of begining date range to plot (yyyy-mm-dd)}

\item{its.end.date.slider}{Character string of ending date range to plot (yyyy-mm-dd)}

\item{its.show.q}{Logical to plot instantaneous hydrograph}

\item{its.show.smooth}{Logical to add a loess smooth to plot}

\item{its.plotpam}{A character vector of parameters to plot}
}
\description{
Takes output data object from NWISPullR and prints a plot of parameter vs time.
}

