% Generated by roxygen2 (4.0.2): do not edit by hand
\name{matrixplot}
\alias{matrixplot}
\title{Parameter matrix plot}
\usage{
matrixplot(qw.data, matrix.site.selection, matrix.plotparm)
}
\arguments{
\item{qw.data}{A qw.data object generated from NWISPullR}

\item{matrix.site.selection}{A character vector of site IDs to plot}

\item{matrix.plotparm}{A character vector of parameters to plot}
}
\description{
Takes output data object from NWISPullR and prints a parameter matrix plot
}

