% Generated by roxygen2 (4.0.2): do not edit by hand
\name{NWISPullR}
\alias{NWISPullR}
\title{Function to pull data from NWIS}
\usage{
NWISPullR(DSN, env.db = "01", qa.db = "02", STAIDS, dl.parms,
  parm.group.check = FALSE, begin.date, end.date)
}
\arguments{
\item{DSN}{A character string containing the DSN for your local server}

\item{env.db}{A character string containing the database number of environmental samples}

\item{qa.db}{A character string containing the database number of QA samples}

\item{STAIDS}{A character vector of stations IDs to pull data}

\item{dl.parms}{A character vector of pcodes to pull data}

\item{parm.group.check}{A logical of weather or not to use NWIS parameter groups. If TRUE, must use NWIS parameter group names in dl.parms}

\item{begin.date}{Character string containing beginning date of data pull (yyyy-mm-dd)}

\item{end.date}{Character string containing ending date of data pull (yyyy-mm-dd)}
}
\description{
Pulls data from NWIS internal servers using ODBC connection and returns a qw.data object.
}

