% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nwisupload}
\alias{nwisupload}
\title{NWISUploader formats files for upload into QWData}
\usage{
nwisupload(qwsampletype, DSN = "NWISCO", labfile, pcodefile,
  qwsamplefile = "", qwresultname = "qwresult", qwsamplename = "qwsample",
  qwsample.begin.date, qwsample.end.date, censor = FALSE,
  agencycode = "USGS", labid = NA, projectcode = NA, aquifercode = NA,
  analysisstatus = NA, analysissource = NA, hydrologiccond = "A",
  hydrologicevent = 9, tissue = NA, bodypart = NA, labcomment = NA,
  fieldcomment = NA, timedatum = NA, datumreliability = NA,
  colagencycode = "USGS")
}
\arguments{
\item{qwsampletype}{Numeric of type of qwsample file. Options 1: Pull qwsample form NWIS, 2:QWSample from file, 3: batch qwsample file}

\item{DSN}{Character string containing local NWIS server name}

\item{labfile}{Character string containing path to the labfile, must be .csv format}

\item{pcodefile}{Character string containing path to the pcodefile, must be .csv format}

\item{qwsamplefile}{Character string containing path to the qwsample file if using qwsampletype = 2, must be .csv format}

\item{qwresultname}{Character string containing output path for the qwresult file}

\item{qwsamplename}{Character string containing output path for the qwsample file}

\item{qwsample.begin.date}{Character string (yyyy-mm-dd) of beggining date to subset qwsample file to}

\item{qwsample.end.date}{Character string (yyyy-mm-dd) of ending date to subset qwsample file to}

\item{censor}{Logical. If TRUE results are censored to lrl provided in pcode file}
}
\description{
Assists in formatting data for upload into QWData. Use Excell spreadsheet templates to create labfile and pcode file.
}

